+++ core/src/main/java/org/mobilitydata/gtfsvalidator/notice/InvalidTimeNotice.java
@@ -21,7 +21,7 @@ import com.google.common.collect.ImmutableMap;
 /**
  * A field cannot be parsed as time.
  *
- * <p>Time must be in the {@code HH:MM:SS} or {@code H:MM:SS} format.
+ * <p>Time must be in the {@code H:MM:SS}, {@code HH:MM:SS} or {@code HHH:MM:SS} format.
  *
  * <p>Example: {@code 14:30:00} for 2:30PM or {@code 25:35:00} for 1:35AM on the next day.
  */
