+++ core/src/main/java/org/mobilitydata/gtfsvalidator/notice/InvalidUrlNotice.java
@@ -22,8 +22,7 @@ import com.google.common.collect.ImmutableMap;
  * A field contains a malformed URL.
  *
  * <p>Definitions for valid URLs are quite vague. We perform strict validation in the upstream using
- * the Apache Common UrlValidator but we allow GTFS consumers to patch the severity level of that
- * validation.
+ * the Apache Common UrlValidator.
  *
  * <p><a href="https://github.com/google/transit/blob/master/gtfs/spec/en/reference.md">GTFS
  * reference</a> describes the requirements in the following way.
